Move omerocmd custom role to the common configuration file
